Metric: Log extended smp pairing events
Include transport and complete information in extended enums.

Tag: #refactor
Bug: 263323082
Bug: 236677941
Test: atest gd/cert/run
(cherry picked from commit 2f74a3c29e801cced2160be3b55266f167b3b202)
Merged-In: I6e527d8ad8b9dfa91b8abf67a9faa0fb11a513fb

+++ system/stack/metrics/stack_metrics_logging.cc
@@ -42,9 +42,9 @@ void log_link_layer_connection_event(
       hci_ble_event, cmd_status, reason_code);
 }
 
-void log_smp_pairing_event(const RawAddress& address, uint8_t smp_cmd,
+void log_smp_pairing_event(const RawAddress& address, uint16_t smp_cmd,
                            android::bluetooth::DirectionEnum direction,
-                           uint8_t smp_fail_reason) {
+                           uint16_t smp_fail_reason) {
   bluetooth::shim::LogMetricSmpPairingEvent(address, smp_cmd, direction,
                                             smp_fail_reason);
 }
